Update to sbt 2.0.0 and cross-publish plugin for sbt 1.x and 2.x
